pyproject.toml
[project]
name = "clipboardToPhone"
version = "0.1.0"
description = ""
authors = [
    {name = "Dist Bit", email = "adp01@pm.me"},
]
dependencies = [
    "python-dotenv>=1.0.1,<2.0.0",
    "loguru>=0.7.2,<0.8.0",
    "pysnooper>=1.2.0,<2.0.0",
    "pyperclip>=1.9.0",
    "requests>=2.32.3",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.1",
    "openai>=1.108.1",
    "python-dateutil>=2.9.0.post0",
    "telethon>=1.37.0",
    "yt-dlp>=2024.11.4",
    "html2text>=2024.2.26",
    "bs4>=0.0.2",
    "eldar>=0.0.8",
    "unidecode>=1.3.8",
    "matplotlib>=3.9.2",
    "urlexpander>=0.0.37",
    "pypdf2",
    "soundcloud-lib>=0.6.1",
    "pytweettoolkit>=1.0.1",
    "websocket-client>=1.8.0",
    "ipfs-cid>=1.0.0",
    "youtube-transcript-api>=1.1.1",
    "readability-lxml>=0.8.4.1",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "mypy>=1.19.1",
    "markdown>=3.10.1",
    "boto3>=1.42.37",
]
requires-python = ">=3.13"
license = {text = "UNLICENCE"}
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
